refactor: PipelineTrainer accepts biome concepts instead of generic parameters
